Adding Android improvements and tests
