+++ packages/amplify_api/android/src/main/kotlin/com/amazonaws/amplify/amplify_api/AmplifyApiPlugin.kt
@@ -38,7 +38,6 @@ class AmplifyApiPlugin : FlutterPlugin, MethodCallHandler {
 
     private lateinit var channel: MethodChannel
     private lateinit var context: Context
-    private val handler = Handler(Looper.getMainLooper())
     private val LOG = Amplify.Logging.forNamespace("amplify:flutter:api")
 
     override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
@@ -52,128 +51,11 @@ class AmplifyApiPlugin : FlutterPlugin, MethodCallHandler {
     override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
         when (call.method) {
             "query" ->
-                query(result, call.arguments as Map<String, Any>)
+                FlutterGraphQLApiModule.query(result, call.arguments as Map<String, Any>)
             "mutate" ->
-                mutate(result, call.arguments as Map<String, Any>)
+                FlutterGraphQLApiModule.mutate(result, call.arguments as Map<String, Any>)
             else -> result.notImplemented()
         }
-    }
-
-    @VisibleForTesting
-    fun query(flutterResult: Result, request: Map<String, Any>) {
-        var document: String
-        var variables: Map<String, Any>
-
-        try {
-            document = request["document"] as String
-            variables = request["variables"] as Map<String, Any>
-        } catch (e: ClassCastException) {
-            createFlutterError(
-                    flutterResult,
-                    FlutterApiErrorMessage.ERROR_CASTING_INPUT_IN_PLATFORM_CODE.toString(),
-                    createErrorMap(e))
-            return
-        } catch (e: Exception) {
-            createFlutterError(
-                    flutterResult,
-                    FlutterApiErrorMessage.AMPLIFY_REQUEST_MALFORMED.toString(),
-                    createErrorMap(e))
-            return
-        }
-        Amplify.API.query(
-                SimpleGraphQLRequest<String>(
-                        document,
-                        variables,
-                        String::class.java,
-                        GsonVariablesSerializer()
-                ),
-                { response ->
-                    var result: Map<String, Any> = mapOf(
-                            "data" to response.data,
-                            "errors" to response.errors.map { it.message }
-                    )
-                    LOG.info("GraphQL query operation succeeded with response: $result")
-                    handler.post { flutterResult.success(result) }
-                },
-                {
-                    LOG.error("GraphQL query operation failed", it)
-                    createFlutterError(
-                            flutterResult,
-                            FlutterApiErrorMessage.AMPLIFY_API_QUERY_FAILED.toString(),
-                            createErrorMap(it))
-                }
-        )
-    }
-
-    @VisibleForTesting
-    fun mutate(flutterResult: Result, request: Map<String, Any>) {
-        var document: String
-        var variables: Map<String, Any>
-
-        try {
-            document = request["document"] as String
-            variables = request["variables"] as Map<String, Any>
-        } catch (e: ClassCastException) {
-            createFlutterError(
-                    flutterResult,
-                    FlutterApiErrorMessage.ERROR_CASTING_INPUT_IN_PLATFORM_CODE.toString(),
-                    createErrorMap(e))
-            return
-        } catch (e: Exception) {
-            createFlutterError(
-                    flutterResult,
-                    FlutterApiErrorMessage.AMPLIFY_REQUEST_MALFORMED.toString(),
-                    createErrorMap(e))
-            return
-        }
-
-        Amplify.API.mutate(
-                SimpleGraphQLRequest<String>(
-                        document,
-                        variables,
-                        String::class.java,
-                        GsonVariablesSerializer()
-                ),
-                { response ->
-                    var result: Map<String, Any> = mapOf(
-                            "data" to response.data,
-                            "errors" to response.errors.map { it.message }
-                    )
-                    LOG.info("GraphQL mutate operation succeeded with response : $result")
-                    handler.post { flutterResult.success(result) }
-                },
-                {
-                    LOG.error("GraphQL mutate operation failed", it)
-                    createFlutterError(
-                            flutterResult,
-                            FlutterApiErrorMessage.AMPLIFY_API_MUTATE_FAILED.toString(),
-                            createErrorMap(it))
-                }
-        )
-    }
-
-    private fun createErrorMap(@NonNull error: Exception): Map<String, Any> {
-        var errorMap = HashMap<String, Any>()
-
-        var localizedError = ""
-        var recoverySuggestion = ""
-        if (error is ApiException) {
-            recoverySuggestion = error.recoverySuggestion
-        }
-        if (error.localizedMessage != null) {
-            localizedError = error.localizedMessage
-        }
-        errorMap["PLATFORM_EXCEPTIONS"] = mapOf(
-                "platform" to "Android",
-                "localizedErrorMessage" to localizedError,
-                "recoverySuggestion" to recoverySuggestion,
-                "errorString" to error.toString()
-        )
-        return errorMap
-    }
-
-    private fun createFlutterError(flutterResult: Result, msg: String, errorMap: Map<String, Any>) {
-        handler.post { flutterResult.error("AmplifyException", msg, errorMap) }
     }
 
     override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
